+++ src/main/java/com/ch/entity/SysUserExample.java
@@ -120,7 +120,7 @@ public class SysUserExample {
             return (Criteria) this;
         }
 
-        public Criteria andUserIdNotEqualTo(Integer value) {
+        public Criteria andUserIdNotEqualTo(Long value) {
             addCriterion("user_id <>", value, "userId");
             return (Criteria) this;
         }
